Done some changes

+++ src/main/java/com/driver/services/impl/ParkingLotServiceImpl.java
@@ -20,38 +20,46 @@ public class ParkingLotServiceImpl implements ParkingLotService {
     SpotRepository spotRepository1;
     @Override
     public ParkingLot addParkingLot(String name, String address) {
-        ParkingLot parkingLot = new ParkingLot(name, address);
-        parkingLotRepository1.save(parkingLot);
+        ParkingLot parkingLot=new ParkingLot();
+        parkingLot.setName(name);
+        parkingLot.setAddress(address);
+        parkingLot=parkingLotRepository1.save(parkingLot);
         return parkingLot;
     }
 
     @Override
     public Spot addSpot(int parkingLotId, Integer numberOfWheels, Integer pricePerHour) {
-        SpotType spotType = null;
-        if(numberOfWheels==2)
-            spotType = SpotType.TWO_WHEELER;
-        else if (numberOfWheels==4) {
-            spotType = SpotType.FOUR_WHEELER;
+        Spot spot=new Spot();
+        spot.setPricePerHour(pricePerHour);
+        spot.setOccupied(Boolean.FALSE);
+        if(numberOfWheels<=2)
+        {
+            spot.setSpotType(SpotType.TWO_WHEELER);
+        }
+        else if(numberOfWheels<=4)
+        {
+            spot.setSpotType(SpotType.FOUR_WHEELER);
         }
         else
-            spotType = SpotType.OTHERS;
-
-        ParkingLot parkingLot = parkingLotRepository1.findById(parkingLotId).get();
-        Spot spot = new Spot(spotType, pricePerHour, false, parkingLot);
+        {
+            spot.setSpotType(SpotType.OTHERS);
+        }
+        ParkingLot parkingLot=parkingLotRepository1.findById(parkingLotId).get();
+        spot.setParkingLot(parkingLot);
         parkingLot.getSpotList().add(spot);
-//        spotRepository1.save(spot);
         parkingLotRepository1.save(parkingLot);
-        return spot;
+        return  spot;
     }
 
     @Override
     public void deleteSpot(int spotId) {
+        if(spotRepository1.existsById(spotId)==false) return ;
         spotRepository1.deleteById(spotId);
     }
 
     @Override
     public Spot updateSpot(int parkingLotId, int spotId, int pricePerHour) {
-        ParkingLot parkingLot = parkingLotRepository1.findById(parkingLotId).get();
+        ParkingLot parkingLot= parkingLotRepository1.findById(parkingLotId).get();
         Spot spot = null;
         for(Spot spot1:parkingLot.getSpotList()){
             if(spot1.getId() == spotId){
@@ -62,10 +70,12 @@ public class ParkingLotServiceImpl implements ParkingLotService {
             }
         }
         return spot;
+
     }
 
     @Override
     public void deleteParkingLot(int parkingLotId) {
+        //if(parkingLotRepository1.existsById(parkingLotId)==false) return ;
         parkingLotRepository1.deleteById(parkingLotId);
     }
 }
